Change of type after using a function.
The function in question being onvm_stats_clear_client.

+++ onvm/onvm_mgr/onvm_stats.c
@@ -76,7 +76,7 @@ onvm_stats_clear_all_clients(void) {
 }
 
 void
-onvm_stats_clear_client(unsigned id) {
+onvm_stats_clear_client(uint16_t id) {
         clients[id].stats.rx = clients[id].stats.rx_drop = 0;
         clients[id].stats.act_drop = clients[id].stats.act_tonf = 0;
         clients[id].stats.act_next = clients[id].stats.act_out = 0;
